api/requirements.txt
# librerias para el endpoint de correo 
fastapi
uvicorn 
dnspython 
pydantic
aiosmtplib
# primer endpoint
tenacity
cfdiclient